Added function to get the state timer

+++ test/Framework.cpp
@@ -27,6 +27,6 @@
 unsigned long emb::Framework::getMilliseconds() {
 
     using namespace std::chrono;
-    return duration_cast<milliseconds>(system_clock::now().time_since_epoch()).count();
+    return (unsigned long) duration_cast<milliseconds>(system_clock::now().time_since_epoch()).count();
 
 }
